Agregando validaciones para que cada usuario puede administrar sus listas independientes. 	modificado:     todo/auth.py 	modificado:     todo/todo.py

+++ todo/todo.py
@@ -12,7 +12,8 @@ bp = Blueprint('todo',__name__)
 def index():
   db,c =get_db()
   c.execute(
-    'select t.id,t.description,u.username,t.completed,t.created_at from todo t join user u on t.created_by = u.id order by created_at desc'
+    'select t.id,t.description,u.username,t.completed,t.created_at from todo t '
+    'join user u on t.created_by = u.id and t.created_by = %s order by created_at desc', (g.user['id'],)
   )
   todos = c.fetchall()
   return render_template('todo/index.html',todos=todos)
@@ -65,7 +66,7 @@ def update(id):
       db,c = get_db()
       c.execute(
         'update todo set description = %s, completed = %s'
-        ' where id = %s', (description,completed,id)
+        ' where id = %s and created_by = %s', (description,completed,id,g.user['id'])
       )
       db.commit()
       return redirect(url_for('todo.index'))
@@ -76,7 +77,7 @@ def update(id):
 def delete(id):
   db,c = get_db()
   c.execute(
-    'delete from todo where id = %s' , (id,)
+    'delete from todo where id = %s and created_by = %s' , (id,g.user['id'])
   )
   db.commit()
   return redirect(url_for('todo.index'))
